# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: MobiCare Pharmacy Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hq/pharmacies:
    get:
      operationId: listHqPharmacies
      tags: [hq]
      summary: All pharmacies
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Pharmacies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HqPharmacy"
    post:
      operationId: onboardPharmacy
      tags: [hq]
      summary: Onboard a pharmacy (returns one-time temp password)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacyOnboardInput"
      responses:
        "201":
          description: Created pharmacy with temp password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyOnboardResponse"
        "409":
          description: Username taken
  /hq/pharmacies/{id}:
    patch:
      operationId: updateHqPharmacy
      tags: [hq]
      summary: Licence verification, Tier-1 gate, online/offline status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HqPharmacyUpdate"
      responses:
        "200":
          description: Updated pharmacy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HqPharmacy"
  /hq/pharmacies/{id}/reset-password:
    post:
      operationId: resetPharmacyPassword
      tags: [hq]
      summary: Generate a new temporary password for an active pharmacy (returns plaintext once)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
      responses:
        "400":
          description: Invalid pharmacy id
        "200":
          description: Temporary password, expiry, and updated pharmacy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacyResetPasswordResponse"
        "404":
          description: Pharmacy not found
        "409":
          description: Pharmacy is inactive
...
  /hq/settlements:
    get:
      operationId: listSettlements
      tags: [hq]
      summary: Pharmacy and courier settlements
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Settlements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementsResponse"
  /hq/settlements/generate:
    post:
      operationId: generateSettlements
      tags: [hq]
      summary: Generate pending settlements for a period
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSettlementsInput"
      responses:
        "201":
          description: Settlements generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /hq/settlements/{id}/mark-paid:
    post:
      operationId: markSettlementPaid
      tags: [hq]
      summary: Mark a settlement as paid
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkPaidInput"
      responses:
        "200":
          description: Paid settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "409":
          description: Already paid
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth schemas ───────────────────────────────────────────────────────
...
    HqPharmacy:
      type: object
      required:
        [
          id,
          name,
          username,
          isActive,
          controlledSubstanceAuthorized,
          mustChangePassword,
          sessionVersion,
          passwordLastChangedAt,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        locationLat:
          type: ["string", "null"]
        locationLng:
          type: ["string", "null"]
        isActive:
          type: boolean
        controlledSubstanceAuthorized:
          type: boolean
        mustChangePassword:
          type: boolean
        sessionVersion:
          type: number
        passwordLastChangedAt:
          type: string
        temporaryPasswordExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PharmacyOnboardInput:
      type: object
      required: [name, username]
      properties:
        name:
          type: string
          minLength: 1
        username:
          type: string
          minLength: 3
        phone:
          type: string
        address:
          type: string
    PharmacyOnboardResponse:
      type: object
      required: [pharmacy, tempPassword, temporaryPasswordExpiresAt]
      properties:
        pharmacy:
          $ref: "#/components/schemas/HqPharmacy"
        tempPassword:
          type: string
          description: Shown once — store it securely and hand it to the pharmacy
        temporaryPasswordExpiresAt:
          type: string
          description: ISO timestamp when the temporary password expires
    PharmacyResetPasswordResponse:
      type: object
      required: [tempPassword, temporaryPasswordExpiresAt, pharmacy]
      properties:
        tempPassword:
          type: string
          description: Shown once — hand it to the pharmacy immediately; it is not stored in plaintext
        temporaryPasswordExpiresAt:
          type: string
          description: ISO timestamp when the temporary password expires
        pharmacy:
          $ref: "#/components/schemas/HqPharmacy"
...
    HqPharmacyUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        controlledSubstanceAuthorized:
          type: boolean
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
...
    Settlement:
      type: object
      required: [id, amountLeones, periodStart, periodEnd, status, createdAt]
      properties:
        id:
          type: string
        pharmacyId:
          type: ["string", "null"]
        pharmacyName:
          type: ["string", "null"]
        courierId:
          type: ["string", "null"]
        courierName:
          type: ["string", "null"]
        amountLeones:
          type: number
        periodStart:
          type: string
        periodEnd:
          type: string
        orderCount:
          type: ["number", "null"]
        deliveryCount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, paid]
        paidAt:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
    SettlementsResponse:
      type: object
      required: [pharmacy, courier]
      properties:
        pharmacy:
          type: array
          items:
            $ref: "#/components/schemas/Settlement"
        courier:
          type: array
          items:
            $ref: "#/components/schemas/Settlement"
    GenerateSettlementsInput:
      type: object
      required: [periodStart, periodEnd]
      properties:
        periodStart:
          type: string
          description: ISO date-time (inclusive)
        periodEnd:
          type: string
          description: ISO date-time (exclusive)
    MarkPaidInput:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [pharmacy, courier]
        reference:
          type: string
...